+++ .kessel/workflows/cmake.py
@@ -63,6 +63,7 @@ class Build(BuildEnvironment):
         """Testing"""
         self.exec(f"""
             pushd {self.build_dir}
+            source "$KESSEL_BUILD_ENV"
             if [[ -f sesame2spiner/sesame2spiner ]]; then
                 echo "/usr/projects/data/eos/eos-developmental/Sn2162/v01/sn2162-v01.bin" > sesameFilesDir.txt
                 sesame2spiner/sesame2spiner -s materials.sp5 ../sesame2spiner/examples/unit_tests/mats.dat